Across Slope calculation updated
Across slope calculation updated to reflect that moving uphill/downhill still has a cost. Previously, if the across slope was greater than 90 degrees then the cost would be 0. However, as the across slope value is a multiplier to the slope, the resultant cost would be 0. This doesn't reflect that slope still has a cost. Updated to return a cost value of 1 when across slope is greater than 90 degrees.